refactor: migrate task
