+++ src/hooks/useNotifications.ts
@@ -1,17 +1,21 @@
-import * as Notifications from 'expo-notifications';
+import notifee, { EventType } from '@notifee/react-native';
 import { useEffect } from 'react';
 
-import { INotificationData } from '../utils/notifications';
+import { INotificationData } from '../notifications/types';
 
 const useNotification = (callback: (data: INotificationData) => void) => {
-  const lastNotificationResponse = Notifications.useLastNotificationResponse();
-
   useEffect(() => {
-    if (!lastNotificationResponse) return;
+    notifee.getInitialNotification().then((notification) => {
+      if (!notification) return;
 
-    const data = lastNotificationResponse.notification.request.content.data as INotificationData;
-    callback(data);
-  }, [lastNotificationResponse]);
+      callback(notification.notification.data as unknown as INotificationData);
+    });
+
+    return notifee.onForegroundEvent(async (event) => {
+      if (event.type !== EventType.PRESS) return;
+      callback(event.detail.notification.data as unknown as INotificationData);
+    });
+  }, []);
 };
 
 export default useNotification;
