+++ src/notifications/handler.ts
@@ -3,6 +3,9 @@ import notifee, { Event, EventType } from '@notifee/react-native';
 import { openFile } from '../utils/files';
 import { INotificationData } from './types';
 
+// Если для уведомления не требуется доступ к UI приложения,
+// то их обработка должна быть произведена здесь
+
 const handleEvent = async (event: Event) => {
   if (event.type !== EventType.PRESS) return;
